+++ src/Termonad/App.hs
@@ -5,7 +5,7 @@ module Termonad.App where
 import Termonad.Prelude
 
 import Config.Dyre (defaultParams, projectName, realMain, showError, wrapMain)
-import Control.Lens ((&), (.~), (^.), (^..), over, set)
+import Control.Lens ((&), (.~), (^.), (^..), over, set, view)
 import Data.FocusList (focusList, moveFromToFL, updateFocusFL)
 import Data.Sequence (findIndexR)
 import GI.Gdk (castTo, managedForeignPtr, screenGetDefault)
@@ -29,6 +29,7 @@ import GI.Gtk
   , Dialog(Dialog)
   , Entry(Entry)
   , FontButton(FontButton)
+  , Label(Label)
   , PolicyType(PolicyTypeAutomatic)
   , PositionType(PositionTypeRight)
   , ResponseType(ResponseTypeAccept, ResponseTypeNo, ResponseTypeYes)
@@ -85,6 +86,7 @@ import GI.Gtk
   , widgetDestroy
   , widgetGrabFocus
   , widgetSetCanFocus
+  , widgetSetVisible
   , widgetShow
   , widgetShowAll
   , windowPresent
@@ -119,6 +121,8 @@ import GI.Vte
   , terminalSetScrollbackLines
   , terminalSetWordCharExceptions
   )
+import System.Environment (getExecutablePath)
+import System.FilePath (takeFileName)
 
 import Paths_termonad (getDataFileName)
 import Termonad.Gtk (appNew, objFromBuildUnsafe)
@@ -144,6 +148,7 @@ import Termonad.Lenses
   , lensTerm
   , lensWordCharExceptions
   )
+import Termonad.PreferencesFile (saveToPreferencesFile)
 import Termonad.Term
   ( createTerm
   , relabelTabs
@@ -758,6 +763,13 @@ showPreferencesDialog mvarTMState = do
     objFromBuildUnsafe preferencesBuilder "scrollbackLen" SpinButton
   adjustmentNew 0 0 (fromIntegral (maxBound :: Int)) 1 10 0 >>=
     spinButtonSetAdjustment scrollbackLenSpinButton
+  warningLabel <- objFromBuildUnsafe preferencesBuilder "warning" Label
+
+  -- We show the warning label only if the user has launched termonad with a
+  -- termonad.hs file
+  executablePath <- getExecutablePath
+  let hasTermonadHs = takeFileName executablePath == "termonad-linux-x86_64"
+  widgetSetVisible warningLabel hasTermonadHs
 
   -- Make the dialog modal
   maybeWin <- applicationGetActiveWindow app
@@ -812,6 +824,9 @@ showPreferencesDialog mvarTMState = do
         . over lensShowTabBar (`fromMaybe` maybeShowTabBar)
         . over lensCursorBlinkMode (`fromMaybe` maybeCursorBlinkMode)
         )
+
+    -- Save the changes to the preferences files
+    withMVar mvarTMState $ saveToPreferencesFile . view lensTMStateConfig
 
     -- Update the app with new settings
     applyNewPreferences mvarTMState