+++ src/Termonad/Types.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE StandaloneDeriving #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Termonad.Types where
 
@@ -6,6 +6,12 @@ import Termonad.Prelude
 
 import Data.FocusList (FocusList, emptyFL, singletonFL, getFocusItemFL, lengthFL)
 import Data.Unique (Unique, hashUnique, newUnique)
+import Data.Yaml
+  ( FromJSON(parseJSON)
+  , ToJSON(toJSON)
+  , Value(String)
+  , withText
+  )
 import GI.Gtk
   ( Application
   , ApplicationWindow
@@ -19,7 +25,7 @@ import GI.Gtk
   , notebookGetNPages
   )
 import GI.Pango (FontDescription)
-import GI.Vte (Terminal, CursorBlinkMode(CursorBlinkModeOn))
+import GI.Vte (Terminal, CursorBlinkMode(..))
 import Text.Pretty.Simple (pPrint)
 import Text.Show (Show(showsPrec), ShowS, showParen, showString)
 
@@ -258,7 +264,7 @@ data FontSize
     -- can be thought of as one pixel.  The function
     -- 'GI.Pango.fontDescriptionSetAbsoluteSize' is used to set the font size.
     -- See the documentation for that function for more info.
-  deriving (Eq, Show)
+  deriving (Eq, FromJSON, Generic, Show, ToJSON)
 
 -- | The default 'FontSize' used if not specified.
 --
@@ -300,7 +306,7 @@ data FontConfig = FontConfig
     -- ^ The font family to use.  Example: @"DejaVu Sans Mono"@ or @"Source Code Pro"@
   , fontSize :: !FontSize
     -- ^ The font size.
-  } deriving (Eq, Show)
+  } deriving (Eq, FromJSON, Generic, Show, ToJSON)
 
 -- | The default 'FontConfig' to use if not specified.
 --
@@ -347,7 +353,7 @@ data ShowScrollbar
                         -- needed.
   | ShowScrollbarIfNeeded -- ^ Only show the scrollbar if there are too many
                           -- lines on the terminal to show all at once.
-  deriving (Enum, Eq, Show)
+  deriving (Enum, Eq, Generic, FromJSON, Show, ToJSON)
 
 -- | Whether or not to show the tab bar for switching tabs.
 data ShowTabBar
@@ -355,7 +361,7 @@ data ShowTabBar
                     -- open.  This may be confusing if you plan on using multiple tabs.
   | ShowTabBarAlways -- ^ Always show the tab bar, even if you only have one tab open.
   | ShowTabBarIfNeeded  -- ^ Only show the tab bar if you have multiple tabs open.
-  deriving (Enum, Eq, Show)
+  deriving (Enum, Eq, Generic, FromJSON, Show, ToJSON)
 
 -- | Configuration options for Termonad.
 --
@@ -386,7 +392,22 @@ data ConfigOptions = ConfigOptions
     -- ^ When to show the tab bar.
   , cursorBlinkMode :: !CursorBlinkMode
     -- ^ How to handle cursor blink.
-  } deriving (Eq, Show)
+  } deriving (Eq, Generic, FromJSON, Show, ToJSON)
+
+instance FromJSON CursorBlinkMode where
+  parseJSON = withText "CursorBlinkMode" $ \c -> do
+    case (c :: Text) of
+      "CursorBlinkModeSystem" -> pure CursorBlinkModeSystem
+      "CursorBlinkModeOn" -> pure CursorBlinkModeOn
+      "CursorBlinkModeOff" -> pure CursorBlinkModeOff
+      _ -> fail "Wrong value for CursorBlinkMode"
+
+instance ToJSON CursorBlinkMode where
+  toJSON CursorBlinkModeSystem = String "CursorBlinkModeSystem"
+  toJSON CursorBlinkModeOn = String "CursorBlinkModeOn"
+  toJSON CursorBlinkModeOff = String "CursorBlinkModeOff"
+  -- Not supposed to happened fall back to system
+  toJSON (AnotherCursorBlinkMode _) = String "CursorBlinkModeSystem"
 
 -- | The default 'ConfigOptions'.
 --